+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+- Use the Node.js `require.cache` for caching the exports returned from a
+  Hook's `onrequire`. This allows users to delete entries from `require.cache`
+  to trigger a re-load (and re-run of the hook's `onrequire`) of a module the
+  next time it is required -- as mentioned at
+  https://nodejs.org/docs/latest/api/all.html#all_modules_requirecache
+  (https://github.com/elastic/require-in-the-middle/issues/61)
+
+- (SEMVER-MAJOR) Remove the `hook.cache` field. In earlier versions this was
+  available and some tests used it. However it was never a documented field.
+
 - If resolving the filename for a `require(...)` fails, defer to the wrapped
   require implementation rather than failing right away. This allows a
   possibly-monkey-patched `require` to do its own special thing.